docs: update README (#461)
The referenced standard README (https://github.com/RichardLitt/standard-readme) says to use a single h1 '#' at the top, and the all subheadings are at most an h2 '##'. This updates the README to follow that guideline, which is also how the docs site is structured. Also, the table of contents was pulling in the main title as well as the "Table of Contents" heading into the actual TOC—this removes them.

For those using the VS Code "Markdown All in One" extension, you can adjust your workspace settings to account for how the TOC is auto-generated. Note that it's possible to extend how many heading levels are displayed in the TOC in case a deeper h4 is desired.
```
"markdown.extension.toc.levels": "2..3",
"markdown.extension.toc.omittedFromToc": {
  "README.md": ["## Table of Contents"]
},
```